# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for RN Codegen Studio
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Whether this is a native module or native component spec
 */
export type GenerateCodeRequestSpecType =
  (typeof GenerateCodeRequestSpecType)[keyof typeof GenerateCodeRequestSpecType];

export const GenerateCodeRequestSpecType = {
  NativeModule: "NativeModule",
  NativeComponent: "NativeComponent",
} as const;

/**
 * Target platform for code generation
 */
export type GenerateCodeRequestPlatform =
  (typeof GenerateCodeRequestPlatform)[keyof typeof GenerateCodeRequestPlatform];

export const GenerateCodeRequestPlatform = {
  android: "android",
  ios: "ios",
  both: "both",
} as const;

export interface GenerateCodeRequest {
  /** Whether this is a native module or native component spec */
  specType: GenerateCodeRequestSpecType;
  /** Target platform for code generation */
  platform: GenerateCodeRequestPlatform;
  /** TypeScript spec content */
  spec: string;
  /** Name of the module or component */
  moduleName: string;
}

export type GenerateCodeResponseSpecType =
  (typeof GenerateCodeResponseSpecType)[keyof typeof GenerateCodeResponseSpecType];

export const GenerateCodeResponseSpecType = {
  NativeModule: "NativeModule",
  NativeComponent: "NativeComponent",
} as const;

export type GenerateCodeResponsePlatform =
  (typeof GenerateCodeResponsePlatform)[keyof typeof GenerateCodeResponsePlatform];

export const GenerateCodeResponsePlatform = {
  android: "android",
  ios: "ios",
  both: "both",
} as const;

export type GeneratedFileLanguage =
  (typeof GeneratedFileLanguage)[keyof typeof GeneratedFileLanguage];

export const GeneratedFileLanguage = {
  java: "java",
  kotlin: "kotlin",
  cpp: "cpp",
  objc: "objc",
  swift: "swift",
  typescript: "typescript",
} as const;

export type GeneratedFilePlatform =
  (typeof GeneratedFilePlatform)[keyof typeof GeneratedFilePlatform];

export const GeneratedFilePlatform = {
  android: "android",
  ios: "ios",
  shared: "shared",
} as const;

export interface GeneratedFile {
  filename: string;
  language: GeneratedFileLanguage;
  platform: GeneratedFilePlatform;
  content: string;
}

export interface GenerateCodeResponse {
  id: number;
  moduleName: string;
  specType: GenerateCodeResponseSpecType;
  platform: GenerateCodeResponsePlatform;
  files: GeneratedFile[];
  generatedAt: string;
}

export type TemplateSpecType =
  (typeof TemplateSpecType)[keyof typeof TemplateSpecType];

export const TemplateSpecType = {
  NativeModule: "NativeModule",
  NativeComponent: "NativeComponent",
} as const;

export interface Template {
  id: string;
  name: string;
  description: string;
  specType: TemplateSpecType;
  moduleName: string;
  spec: string;
}

export interface TemplateListResponse {
  templates: Template[];
}

export type HistoryItemSpecType =
  (typeof HistoryItemSpecType)[keyof typeof HistoryItemSpecType];

export const HistoryItemSpecType = {
  NativeModule: "NativeModule",
  NativeComponent: "NativeComponent",
} as const;

export type HistoryItemPlatform =
  (typeof HistoryItemPlatform)[keyof typeof HistoryItemPlatform];

export const HistoryItemPlatform = {
  android: "android",
  ios: "ios",
  both: "both",
} as const;

export interface HistoryItem {
  id: number;
  moduleName: string;
  specType: HistoryItemSpecType;
  platform: HistoryItemPlatform;
  generatedAt: string;
  fileCount: number;
}

export interface HistoryListResponse {
  items: HistoryItem[];
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
  details?: string;
}