# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for RN Codegen Studio
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Generates platform-specific native code from a TypeScript spec
 * @summary Generate native code from spec
 */
export const GenerateCodeBody = zod.object({
  specType: zod
    .enum(["NativeModule", "NativeComponent"])
    .describe("Whether this is a native module or native component spec"),
  platform: zod
    .enum(["android", "ios", "both"])
    .describe("Target platform for code generation"),
  spec: zod.string().describe("TypeScript spec content"),
  moduleName: zod.string().describe("Name of the module or component"),
});

export const GenerateCodeResponse = zod.object({
  id: zod.number(),
  moduleName: zod.string(),
  specType: zod.enum(["NativeModule", "NativeComponent"]),
  platform: zod.enum(["android", "ios", "both"]),
  files: zod.array(
    zod.object({
      filename: zod.string(),
      language: zod.enum([
        "java",
        "kotlin",
        "cpp",
        "objc",
        "swift",
        "typescript",
      ]),
      platform: zod.enum(["android", "ios", "shared"]),
      content: zod.string(),
    }),
  ),
  generatedAt: zod.date(),
});

/**
 * Returns a list of pre-built spec templates to start from
 * @summary List available spec templates
 */
export const ListTemplatesResponse = zod.object({
  templates: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      description: zod.string(),
      specType: zod.enum(["NativeModule", "NativeComponent"]),
      moduleName: zod.string(),
      spec: zod.string(),
    }),
  ),
});

/**
 * Returns the history of code generation requests
 * @summary List generation history
 */
export const ListHistoryResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.number(),
      moduleName: zod.string(),
      specType: zod.enum(["NativeModule", "NativeComponent"]),
      platform: zod.enum(["android", "ios", "both"]),
      generatedAt: zod.date(),
      fileCount: zod.number(),
    }),
  ),
});

/**
 * Clears all generation history
 * @summary Clear generation history
 */
export const ClearHistoryResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * Returns a single generation history item
 * @summary Get a history item
 */
export const GetHistoryItemParams = zod.object({
  id: zod.coerce.number(),
});

export const GetHistoryItemResponse = zod.object({
  id: zod.number(),
  moduleName: zod.string(),
  specType: zod.enum(["NativeModule", "NativeComponent"]),
  platform: zod.enum(["android", "ios", "both"]),
  generatedAt: zod.date(),
  fileCount: zod.number(),
});