# lib/api-zod/src/generated/types/generateCodeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for RN Codegen Studio
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateCodeRequestPlatform } from "./generateCodeRequestPlatform";
import type { GenerateCodeRequestSpecType } from "./generateCodeRequestSpecType";

export interface GenerateCodeRequest {
  /** Whether this is a native module or native component spec */
  specType: GenerateCodeRequestSpecType;
  /** Target platform for code generation */
  platform: GenerateCodeRequestPlatform;
  /** TypeScript spec content */
  spec: string;
  /** Name of the module or component */
  moduleName: string;
}