# lib/api-zod/src/generated/types/generateCodeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for RN Codegen Studio
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateCodeResponsePlatform } from "./generateCodeResponsePlatform";
import type { GenerateCodeResponseSpecType } from "./generateCodeResponseSpecType";
import type { GeneratedFile } from "./generatedFile";

export interface GenerateCodeResponse {
  id: number;
  moduleName: string;
  specType: GenerateCodeResponseSpecType;
  platform: GenerateCodeResponsePlatform;
  files: GeneratedFile[];
  generatedAt: Date;
}