# lib/api-zod/src/generated/types/generatedFile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for RN Codegen Studio
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratedFileLanguage } from "./generatedFileLanguage";
import type { GeneratedFilePlatform } from "./generatedFilePlatform";

export interface GeneratedFile {
  filename: string;
  language: GeneratedFileLanguage;
  platform: GeneratedFilePlatform;
  content: string;
}