# lib/api-zod/src/generated/types/template.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for RN Codegen Studio
 * OpenAPI spec version: 0.1.0
 */
import type { TemplateSpecType } from "./templateSpecType";

export interface Template {
  id: string;
  name: string;
  description: string;
  specType: TemplateSpecType;
  moduleName: string;
  spec: string;
}